+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Main (unreleased)
 
 ### Enhancements
 
+- Add `rfc3164_default_to_current_year` argument to `loki.source.syslog` (@dehaansa)
+
 - Have `loki.echo` log the `entry_timestamp` and `structured_metadata` for any loki entries received (@dehaansa)
 
 - Update mysqld_exporter to v0.17.2, most notable changes: (@cristiangreco)